+++ run.py
@@ -1,0 +1,6 @@
+from app import create_app
+from app.config import PORT
+
+if __name__ == '__main__':
+    app = create_app()
+    app.run(debug=False, port=PORT)
